Enable containment for some gpu bots
This is (as far as I know) the first time chromium is using containment,
so I'm enabling on a bot which shouldn't affect the CQ that much. Once
this lands, I'll enable on the rest of the CQ bots.

Bug: 965222
Change-Id: Ie0f85f50eec68c4de89f12047209b9ad570ec129
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1626149
Reviewed-by: Aaron Gable <agable@chromium.org>
Reviewed-by: John Budorick <jbudorick@chromium.org>
Reviewed-by: Kenneth Russell <kbr@chromium.org>
Commit-Queue: Stephen Martinis <martiniss@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#662823}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: e314c1166a4daa78c1108a720bad0600c9f4871c