Set up NativeBrowserTestApplications in the base class.
We move the ApplicationStatus.initialize() call up to the base class
since they all need to do it.

We also initialize the CommandLine there as it needs to be initialized
from the Application. The ChromeTabbedActivity would crash during
startup when it tries to use the CommandLine otherwise.

Do the same in NativeTestApplication base class, instead of in
NativeTest, for consistency. This prevents the CommandLine from being
destroyed in NativeTest after the NativeBrowserTestApplication has
already set it up and is running.

Bug: 961849
Change-Id: I715d91fb1b251d4dbbb51e2b60760fb1cc95bc97
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1670349
Reviewed-by: Ted Choc <tedchoc@chromium.org>
Reviewed-by: David Trainor <dtrainor@chromium.org>
Commit-Queue: danakj <danakj@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#671355}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: a8ab3bd5e7aac5fe791620c1bbd2737f5c55825a