Stream goldctl output on V8 builders
Streams the output from running goldctl commands in the GPU pixel tests
in real time on the V8 bots instead of waiting for the command to
finish. This is to help diagnose timeout issues that are only
occurring on the V8 bots.

Bug: 977620
Change-Id: Ibcf6a80e823a4226369f9633eb0e1b477476299a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1680941
Commit-Queue: Brian Sheedy <bsheedy@chromium.org>
Reviewed-by: Kenneth Russell <kbr@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#673105}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 854c62bc2cafab1345a359f7ce6d8db2ccd5c17d