Make DataReductionProxyURLLoaderThrottle work on UI thread
We are in the process of migrating NavigationURLLoaderImpl to the UI
thread, which includes migrating all throttles. This makes the
DataReductionProxyURLLoaderThrottle able to be created on either the UI
or IO thread.

The main changes are giving DataReductionProxyService access to
proxies_for_http so they can be used on the UI thread, and passing
through the mojom::DataReductionProxy calls to
DataReductionProxyIOData.

Bug: 978614
Change-Id: Iaa7be5125e0cdfd26f301467449be871d8a594bf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1678167
Commit-Queue: rajendrant <rajendrant@chromium.org>
Auto-Submit: Clark DuVall <cduvall@chromium.org>
Reviewed-by: rajendrant <rajendrant@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#673232}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 68eb997f28060c9ab0e134075ced56b18e738c18