[Extensions] De-flake MultiActionAPITest.DynamicSetIcon
There is a race condition where, since the default icon for an extension
action loads asynchronously, it may not be ready by the time we check
its value in the test. Fix this by waiting for the default icon to fully
load before checking it against the expected color.

This results in the test passing locally (on an official CrOS build)
100% of the time over 30 runs (where before it would fail regularly).

Bug: 979226
Change-Id: I00b22aee6ef668d6438d41ecf326362640abb022
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1681079
Reviewed-by: Karan Bhatia <karandeepb@chromium.org>
Commit-Queue: Devlin <rdevlin.cronin@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#673812}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 187ae46da27f30b59650a343f4836564af74cd0d