service worker: Initial patch for navigation on the UI thread.
This adds ServiceWorkerNavigationLoaderInterceptor which lives on the
UI thread. Currently it just delegates to the existing interceptor on
the IO thread. The goal is to migrate things to the UI thread.

There are no new tests added as this is still initial work, but
it enables several of the existing browser tests.

Change-Id: I69d9b09c308b0e2c711a304390c04574c734a2d9
Bug: 824858
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1688592
Commit-Queue: Matt Falkenhagen <falken@chromium.org>
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#676354}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 82faa92ed74916208c1f5e2c8312aaaf86465949