[libFuzzer] Docs: re-write the Efficient Fuzzing Guide as per feedback from the tech writer.
Also removed "Custom options" section as it doesn't need to be advertised anymore.
The only real usecase we still may have is `close_fd_mask`, but it's discouraged and
we better use it on a case-by-case basis, when it's inevitable.

Bug: 539572, 827228
Change-Id: I036303e1da0c844be4c3970ba5351651a2251a89
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1729847
Commit-Queue: Max Moroz <mmoroz@chromium.org>
Reviewed-by: Jonathan Metzman <metzman@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#683649}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 4a8415ab474840523f579667bd13a391891ba3ed