Rename Mac backgrounding field trial enabled group
We disabled a stronger version of this a few years ago because of
regressions, with the anticipation that we'd fix some issues and try
again. Since that never happened, I want to look at seeing whether
the weak version is worth launching. This CL updates the field trial
in anticipation of a new Finch config.

Bug: 679417
Change-Id: Ia5e784d03bef0d1cb44da2327a33141c931c3119
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1733748
Reviewed-by: Robert Kaplow <rkaplow@chromium.org>
Commit-Queue: Leonard Grey <lgrey@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#685986}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 2616c2927ce037e67c5b30f32d8401d1e27efbad