Switch GPU Angle perf to histograms
Switches the GPU Angle perf tests to use histograms instead of the
default ChartJSON format. The use of ChartJSON was causing the uploader
script to try to upload to the perf dashboard via the IP whitelisted
endpoint instead of using a service account, which was causing the
upload to fail.

TBR=kbr@chromium.org

Bug: 924618
Change-Id: I7238ff3fe188e3983e91e6391e52fb417caedfd4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1749455
Reviewed-by: Brian Sheedy <bsheedy@chromium.org>
Commit-Queue: Brian Sheedy <bsheedy@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#686089}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 6b30d4e5880e10f6a8cf534c9cb7f5c6e3fc1828