bfcache: Don't record MaxFrameCount for frames in bfcache
The MaxFrameCount UMA records the maximum number of frames in a page,
and is triggered by RenderFrameDeleted. Since RenderFrameDeleted is not
called on a bfcached frame when we navigate away from it (but will be
called when it is evicted from the bfcache/the tab is closed/etc),
the |max_frame_count_| is already overwritten and should not be
recorded.

Bug: 1010870
Change-Id: I669e91aec29af74d3b9e154356b034b50aeaf75a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1855543
Commit-Queue: Rakina Zata Amni <rakina@chromium.org>
Reviewed-by: Alex Moshchuk <alexmos@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#705796}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 74627df81cea6b76c6aa99ed690693adf9309158