bfcache: Update test. Clear in-page memory cache correctly.
To clear the in-memory cache, the test was relying on a cross-process
navigation. This is not enough, because the old page can be kept in the
back-forward cache.

This patch disables the back-forward cache for the old page.

Bug: 999849
Change-Id: I7f2587d0bc25a349c74c8a68599e74db4eb8e318
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1863296
Commit-Queue: Mohamed Abdelhalim <zetamoo@chromium.org>
Reviewed-by: Camille Lamy <clamy@chromium.org>
Reviewed-by: Alexander Timin <altimin@chromium.org>
Reviewed-by: Arthur Sonzogni <arthursonzogni@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#706466}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: d74c8dbc965c41098fd62fc3fcae25e94cdd4935