Turn on by default platform font skia on windows.
The PlatformFontSkiaOnWindows is changing the use of
PlatformFontWin by PlatformFontSkia.
The experiment is deployed on canary and dev.

This CL setting the default state of the experiment to
enroll user in the PlatformFontSkia group.

The legacy code (PlatformFontWin_ is kept as a kill-switch
and will be remove soon.


Bug: 944227
Change-Id: Ib32f2fdee225650d6f0728c2f791d358857f853c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1888314
Reviewed-by: Michael Wasserman <msw@chromium.org>
Reviewed-by: Alexei Svitkine <asvitkine@chromium.org>
Commit-Queue: Etienne Bergeron <etienneb@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#711332}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 03715bcae4db3745f97980d7ff72dcd3ed861f88