Add NavigationPredictorMultiplePrerenders field trial config
Add an entry for NavigationPredictorMultiplePrerenders in field trial
testing config. This maps to the 5 prerender arm in the experiment (the
most aggressive). We are not sure which arm will launch if any, so we
might as well test the most aggressive behavior against the bots.

Bug: 1016213
Change-Id: Ie8e930790b9a4438a32df76496bd7a6a183156ea
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1900469
Reviewed-by: Robert Ogden <robertogden@chromium.org>
Reviewed-by: Brian White <bcwhite@chromium.org>
Reviewed-by: Tarun Bansal <tbansal@chromium.org>
Commit-Queue: Ryan Sturm <ryansturm@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#713332}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: a9473f5884d8fd5ccbaadefd6e94948e05319f7c