Enable xr_browser_tests on Mac
Enables the xr_browser_tests target on Mac bots in addition to the
existing Linux and Windows bots. As a result, also splits the target
into two source sets so that it can be built with or without the
enalbe_vr gn arg specified. This is due to enable_vr being true for
Linux and Windows, but not on Mac.

Bug: 933606
Change-Id: I4f27f51990d8df3417f5826a911e087dbf530cd9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1480546
Commit-Queue: Brian Sheedy <bsheedy@chromium.org>
Reviewed-by: John Budorick <jbudorick@chromium.org>
Reviewed-by: Ben Pastene <bpastene@chromium.org>
Reviewed-by: Alexander Cooper <alcooper@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#715994}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: d9865b8e60ea84ff6a314137ac57eac11926abff